requirements.txt
ouimeaux
paho-mqtt
